Upadte API call structure.
