+++ src/components/EFTs.js
@@ -1,17 +1,8 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { getEFTs } from '../redux/actions/markets';
+import React from 'react';
+import { useSelector } from 'react-redux';
 
 function EFTs() {
   const efts = useSelector((state) => state.efts);
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (!efts.length) {
-      dispatch(getEFTs());
-    }
-  }, []);
 
   return (
     <section className="efts-page">
