+++ src/components/Home.js
@@ -2,22 +2,34 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { categories } from './data';
-import { getForex, getMajors } from '../redux/actions/markets';
+import {
+  getForex, getCryptos, getStocks, getEFTs,
+} from '../redux/actions/markets';
 import '../styles/Home.css';
 
 function Home() {
   const forex = useSelector((state) => state.forex);
-  const majors = useSelector((state) => state.majors);
+  const cryptos = useSelector((state) => state.cryptos);
+  const stocks = useSelector((state) => state.stocks);
+  const efts = useSelector((state) => state.efts);
 
   const dispatch = useDispatch();
 
   useEffect(() => {
     if (!forex.length) {
       dispatch(getForex());
-      dispatch(getMajors(forex));
     }
-    if (forex.length && !majors.length) {
-      dispatch(getMajors(forex));
+
+    if (!cryptos.length) {
+      dispatch(getCryptos());
+    }
+
+    if (!stocks.length) {
+      dispatch(getStocks());
+    }
+
+    if (!efts.length) {
+      dispatch(getEFTs());
     }
   }, []);
 
@@ -30,6 +42,7 @@ function Home() {
           <i className="cog icon" />
         </div>
       </div>
+
       <div className="banner home">
         {categories.map((item) => {
           const { path, category, market_cap: cap } = item;
@@ -44,7 +57,9 @@ function Home() {
           );
         })}
       </div>
+
       <h4 className="sub-heading">Popular</h4>
+
       <div className="content-body">
         {categories.map((item) => {
           const { path, category, market_cap: cap } = item;
@@ -60,6 +75,7 @@ function Home() {
           );
         })}
       </div>
+
       <div className="bottom-bar">Bottom</div>
     </section>
   );
