+++ src/components/Stocks.js
@@ -1,17 +1,8 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { getStocks } from '../redux/actions/markets';
+import React from 'react';
+import { useSelector } from 'react-redux';
 
 function Stocks() {
   const stocks = useSelector((state) => state.stocks);
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (!stocks.length) {
-      dispatch(getStocks());
-    }
-  }, []);
 
   return (
     <section className="stocks-page">
